+++ src/modules/users/repositories/IUsersRepository.ts
@@ -1,4 +1,4 @@
-import User from '../infra/typeorm/entities/user';
+import User from '../infra/typeorm/entities/User';
 import ICreateUserDTO from '../dtos/ICreateUserDTO';
 import IFindAllProvidersDTO from '../dtos/IFindAllProvidersDTO';
