+++ src/modules/users/services/AuthenticateUserService.ts
@@ -1,7 +1,7 @@
 import { sign } from 'jsonwebtoken';
 import { injectable, inject } from 'tsyringe';
 import AppError from '@shared/errors/AppError';
-import User from '@modules/users/infra/typeorm/entities/user';
+import User from '@modules/users/infra/typeorm/entities/User';
 import authConfig from '@config/auth';
 import IUserRepository from '../repositories/IUsersRepository';
 import IHashProvider from '../providers/HashProvider/models/IHashProvider';
